Initial implementation of date time math command
Currently it can parse dates and dump them out in different formats. I
had to switch up how the query is passed in to support having - at the
beginning of a word. How I previously was doing it that was treated as
an invalid argument. Will be needed for this later when math is
supported on the date.