Add a dev server that has single endpoint that runs the cli
Using subprocess and `go run` instead of importing the package, so that
the server can be left running and pick up any new changes. The idea is
that this can be used to develop the alfred extension from a codespace.
The devserver would be started at codespace start and then forward
localhost to the codespace and use curl in the workflow instead of
calling the cli directly.